DB creation issue is fixed

+++ assignment_6.py
@@ -9,11 +9,13 @@ class Assignment6:
     __db = DatabaseUtil(utility.database_name)
 
     def __setup_db(self):
+        if os.path.exists(utility.database_name):
+            os.remove(utility.database_name)
+
         date = utility.read_file("as5.txt", 3)
         temp = utility.read_file("as5.txt", 8)
         hour = utility.read_file("as5.txt", 4)
 
-        os.remove(utility.database_name)
         self.__db.create_db()
 
         i = 0
